+++ CHANGELOG.md
@@ -4,10 +4,9 @@ All notable changes to this project will be documented in this file. See [commit
 
 ## [0.2.71](https://github.com/Wildhoney/Chizu/compare/v0.2.70...v0.2.71) (2026-02-25)
 
-
 ### Features
 
-* added concept of features and made dispatch awaitable ([8049db3](https://github.com/Wildhoney/Chizu/commit/8049db30054b9899572de67b1a53ed4327b0bb67))
+- added concept of features and made dispatch awaitable ([8049db3](https://github.com/Wildhoney/Chizu/commit/8049db30054b9899572de67b1a53ed4327b0bb67))
 
 ## [0.2.70](https://github.com/Wildhoney/Chizu/compare/v0.2.69...v0.2.70) (2026-02-24)